Add max-jobs into chunk morphologies
This lets a morphology specify that a particular chunk should be
run without parallel building. Some projects have build systems that
can't handle parallel building. The default is, of course, to assume
that all projects can handle it. Projects whose build systems do not
allow parallel building are buggy, but we have to accomodate them,
but we won't be making that the default assumption.

This way we gently nudge upstreams to not use broken build systems.