add method to compute recursive cache id
The cache id will be used as a prefix in filenames in the cache
directory. The new way of computing it includes the cache ids of
any components so that if chunk A is included in stratum B, and
A changes, the new version of A has a new cache id, which means
that when we compute cache id for B, it also changes. This will
allow us to easily find out if we need to rebuild B.

The test repository is split into two: one for the chunk, one for
the stratum and system morphologies. This is so that the rebuild
test can modify just the chunk, and does not need to trigger both builds.

Signed-off-by: Lars Wirzenius <liw@liw.fi>