Made some sudo commands into fakeroot commands
Some commands can remain as fakeroot if they share the same fakeroot
This can be done with the -i and -s parameters to fakeroot
Now the Execute objects contain a reference to the file that stores
their state, so fake commands' fake state is recognised by later commands

The creating a system image definitely needs to be sudo though.
Morph needs to create some files as part of the process, so unfortunately
the equivalent to 'echo foo | sudo tee bar >/dev/null' is being used.
They could be created with sufficient permissions, written then have
better permissions set, but it seemed neater this way
A better solution may be to have a subprogram that does the system image
creation

Putting files into the staging area for building other chunks is currently
sudo, but shouldn't have to be. Rather than having a fakeroot per Execute
object, it may be better to have a global one, though this would prevent
concurrent use of fakeroot.