Fix tests to reflect changes
bins now takes an Execute, so the test needed to be changed to also pass one
Execute had to be changed to ignore a statement from coverage
Also fixed Execute.runv not outputting its error. I think it is not logged
this way, but it does get output like it was before.
Also check no longer runs cmdtest from the home directory.