Added a test for cachedir providing absolute paths
Could have used the cachedir to retrieve a name, then os.chdir, then retrieve
with the same dict_key.
However if it could chdir to the cwd the test would behave differently, so
creating a temporary directory would be needed, and would add more code.
So create a Cachedir which explicitly uses an absolute path and compare the
path name returns