Fix Cachedir to convert relative paths to absolute paths
This also required a test to be changed as it assumed that the path
it created the Cachedir with was the same one it stored.
It also technically makes the new test redundant as testing if dirname
is equal to an absolute path also tests if it retrieves values by absolute
path.