Avoid bc as a dependency for baserock-bootstrap
We want to avoid having to have bc in the bootstrap environment,
since it's yet another thing to install there, and to build in
the bootstrap process.

bc is only used to compute the number of parallel make jobs. We can
approximate the answer easily with expr (but need integer arithmetic
so the result is slightly different).