Fix the loop that waits for workers to finish.
Apparently, all(x for x in y) performs some sort of lazy evaluation,
causing it to only check workers to complete until one of them is not
complete. What we want is to check all of them, so we need to use
the all([x for x in y]) notation, which will first generate a list and
then evaluate that by iterating over it. Madness - but it works now!