Cache submodule repositories recursively.
This commit introduces the following new classes:

morphlib.git.Submodules:
  * takes a parent repo treeish
  * parses the .gitmodule file into morphlib.git.Submodule objects
  * provides iterator/container functionality for submodules

morphlib.git.Submodule:
  * represents a single entry in a .gitmodules file
  * stores a Treeish for the corresponding repository

In addition to this, the exception classes InvalidTreeish and
SourceNotFound where renamed to InvalidReferenceError and
RepositoryUpdateError. Several new exception classes were added for when
resolving submodules fails.

The SourceManager now resolves the Submodules and Submodule objects for
the submodules of a Treeish in SourceManager.get_treeish() and also
takes care of caching submodule repositories whenever necessary.