Switch to caching repos without checking them out.
This is almost as good as --bare. But unlike --bare, it properly sets up
the 'origin' remote and creates the 'refs/remotes/origin/...' branches.

When cloning with --bare, and you end up with "refs/heads/master", not
"/refs/remotes/origin/master". You can re-add the remote 'origin', still
no "refs/remotes/origin/master". Only after you update the remote (after
re-adding it), the bare clone has "refs/remotes/origin/master"... and a
"refs/heads/master" that is out of sync. An ugly mix, basically.

So maybe this approach is better.