Delay bundle and base URL errors until we are sure all of them failed.
This is done to not print confusing errors in situations where one of
the bundles or base URLs works and everything is fine. If all of them
are failing though, we now print all the error messages in order to give
users an overview over what might have gone wrong.