builder.py: use sh -c instead of writing a script
Putting the commands in a file made it more difficult to tell
what happened in scripts, until the self.msg was added.
After then it made the commands more difficult to reproduce
So if it can safely be done in a single command it should be more
readable.

This approach uses sh -c to cd then evaluate the remaining options
The command is 'cd "$1" && shift && eval "$@"', then passes the command
as arguments to the shell script, so the program does not need to do any
escaping of strings.

If the commands were a list of strings instead of a quoted string then
it would be a bit safer, but less flexible for use.