Use "cp -a" and "git checkout" to unpack sources in the build tree.
This is done to keep the git repository intact with history and all
that. cp + checkout turns out to be faster than a regular "git clone".

In order to avoid roundtrips to the internet whenever a chunk tries to
do something with submodules, we force the submodule URLs to point to
our locally cached repos.