Hide git output from unit test output
This makes it easier to check the output with vgrep for errors or
problems. Voluminous git output made that hard.

The key is that "git bundle create" has no --quiet optio (meh),
so we have to redirect stderr to /dev/null. However, that would
hide real errors, which we want to catch. Luckily, we can just
replace the call to subprocess.call with a call to
subprocess.check_call, and then we'll at least abort if there's
an error, instead of silently ignoring it.