scripts: don't skip if not snapshotting
run-bootstrap-in-chroot would not build a stage if it had a snapshot
for the next stage.
If the same directory had previously run while snapshotting then
it would skip earlier stages without building them as it used to be
able to extract that tarball.
This fails as stages depend on earlier stages.

This change is so that it will build if it is not snapshotting or
if the stage has not been snapshotted.