morph: drop special case .git in mtime setting
Either all .gits need to be left alone, including submodules, or
all of them get changed.
Doing all gets surprisingly complicated, as bottom-up traversal is
needed for directories to have their mtimes set, which means that
either we check the path for a .git component, or we do our own
traversal function, to top-down exclude .git and bottom-up set
mtimes.

This is more effort than it is worth when git probably doesn't care
so let's just drop it entirely.