Remove the system testing subcommands and related code
This we a proof-of-concept, and we want to explore various options of
how to do this, so let's not keep the code in morph. The code also has
no tests, so it is already likely to have bit-rotted.

If we decide we want to, we can always resurrect it from git history.