Remove all but Python bytecode files from .gitignore
Some like to see tests/*-actual files in "git status" output, for
example, and others don't. The best way to deal with this seems to
be to not have them in .gitignore and add a personal global
ignores file for those who want to ignore them.

git config --global core.excludesfile ~/.git-default-ignores

Then edit ~/.git-default-ignores.