Set sensible defaults for git-base-url, bundle-server, cachedir, max-jobs
Also, change builder.py to always obey the --max-jobs setting, unless
a morphology has a max-jobs field.

The defaults have been chosen so that they work for everyone equally
well. It may be useful to have a local mirror and then set the options
to point there, but it's not reasonable to try to guess such things,
so the defaults can be adapated to that.

Collect the defaults into one place so they're easier to overview.
The cliapp interface for adding settings is verbose enough that the
defaults were getting buried.