Use dd to create an image file, instead of qemu-img
We don't want to have to add all of qemu to Baserock, just for qemu-img.
Since we use a raw disk image file type, we can just create a big file
full of zeroes, and the best way to do that is to create a giant hole.
dd can do that.