Do without kpartx and remove use of sudo
These two changes accidentally got intermingled, oops.

kpartx is a nice tool, but it's not necessary: we can just use
losetup and sfdisk and do things ourselves.

Almost nothing the script does works as a normal user, so it's
not a good idea to sprinkle the code with calls to sudo. Instead,
we'll just require the caller to run it as root.