Exclude from test coverage stuff without tests
The code isn't making it easy to test just the size parsing code.
This code shouldn't even be in morph, it should be re-using a
library (such as the one in cliapp, but cliapp doesn't expose it
nicely). All of this should be fixed, but for now, I'll just
exclude things from test coverage. Shame be on me.