Move pass1_file to the top right after pass1_zlib.
This fixes a strange issue where the file binary will segfault in
__libc_start_main() because it uses the host runtime dynamic linker
and not the on in /tree/tools. The host's ld-linux comes from eglibc
2.12 and the rtld_global_ro struct in 2.12 has a different size than the
one in eglibc 2.15. This, the _rtld_global_ro.dl_audit pointer gets
corrupted and __libc_start_main() segfaults.

This is miraculously fixed by building file immediately after zlib.