Empty and initialize the test data dir anew for each test
This avoids accidentally having tests that depend on each other. It
would be bad to have tests/bar.script do something and then
tests/foo.script only work because of that something: if tests/bar.script
later changes, then the later test starts breaking. Keeping tests
independent is good practice. We don't seem to have had such
dependencies yet, so this is a pre-emptive defensive move.

Also, this means we can use simpler names in $DATADIR, instead of
embedding the test name in everything. This further simplifies the
act of creating new tests.