morph: hash the toolchain .meta files in ccache
We need the entire toolchain to be part of the hash that determines
ccache objects, as if any of them differ it could generate different
output.

ccache can have the cache-key depend on arbitrary files with the
CCACHE_EXTRAFILES environment variable, which is a : separates
list of files to hash.

So add the .meta files of our toolchain to CCACHE_EXTRAFILES if they
exist.

Extra work needs to be done to ensure the toolchain is properly hashed
however, the .meta files will need the cache-key to be included.