Add a new morphology parser class
The old morphlib.morphology.Morphlogy class is entangled by having
treeish and requires reading from an open file. This is a bad design
for the class: the treeish is unnecessary coupling, and the open file
makes is harder than necessary to parse a morphology which we don't
have as a file on disk.

The new class gets the text of the morphology and does not care about
treeishes at all. It also acts more like a dictionary, giving more
uniform access to the various items, while reducing the amount of
code we have in the class.

The old class will remain until all uses of it have been eradicated.