Add a Source class
This will eventually replace the Blob class we are currently using.
It is somewhat simpler, because the envisioned usage of Source will
be different from Blob: we'll be rewriting the build ordering class,
and the builder classes, in ways that simplify things, and thus a
simpler Source class will hopefully suffice.