Add Artifact and LocalArtifactCache classes.
An Artifact represents a thing that morph has built. An example would
be eglibc-runtime which morph may have built from the eglibc chunk
morphology. Another example would be a ready-to-use system image.

The LocalArtifactCache allows to store build artifacts in a local
directory. Users of this class can ask it whether it has a certain
artifact. They can also optain an I/O handle to read the artifact data
from.

In addition to just abstracting the way artifacts are stored,
LocalArtifactCache also allows to store and retrieve metadata for
(a) artifacts and (b) sources (the latter requires a cache key to
be provided to the LocalArtifactCache).