morphlib: add buildenvironment and tests
BuildEnvironment should contain all the information about whatever
settings affect the build. A subset of these will affect the cache
key, but CacheKeyComputer filters them itself.

BuildEnvironment has consumed the purpose of Morph.clean_env() because
the build environment is a good place to store this information.

The environment variables are decided based on the current settings.
The current environment can be accessed by the env attribute.

The tests are a little invasive because it reads environment variables
and most of the work is done in the constructor, but they should test
everything useful.