Remove BuildGraph, compute cache keys based on Artifacts.
With this commit, the ArtifactResolver no longer computes the cache keys
when creating Artifact objects. This will have to happen as a
post-resolving step (e.g. prior to building or checking whether a local
or remote artifact cache has any of the resolved artifacts).

The CacheKeyComputer now takes an Artifact object and computes the cache
keys using its dependencies.

BuildGraph is no longer needed for the CacheKeyComputer unit tests.