Drop httplib2 in RemoteRepoCache, use custom HeadRequest and urllib2.
With a neat little trick, urllib2 can be made to only perform a HEAD
request in urllib2.urlopen(). This is done by creating a dedicated
HeadRequest class for HEAD requests and by passing such a HeadRequest to
urllib2.urlopen().

This also means we no longer have to check the response status ourselves
again as urllib2 will raise an exception on negative HTTP responses like
404 or 500.