morphologyfactory: infer build system without list
Detecting the build system is managed by it asking if any files exist
detecting if the file exists is done with a callback function.
This callback can use cat-file.

If list_files existed this could be more efficient as it would not
require the files to be read from the remote server and it only needs
to be one round-trip