morphlib/morph2: default commands are None
Builder checks whether commands need to be chosen from the build
system by if the morphology has no commands, the check is whether
the commands are None, rather than an empty list, so that it is
possible to override a build system's commands with an empty list.

However this means that the default of an empty list means that the
build system is ignored and commands aren't run.