system image use systemd on vda1 and build in tmp
Use the tempdir of the staging area instead of the chroot directory
as the basedir for building system images.
It would be better to create it directly in the cache, then rename it
when finished, but commands need a filename and the cache provides an
open file handle.

Change the generated fstab and extlinux.conf to use systemd as the init
and expect the root filesystem to be on vda1 instead of sda1.
vda1 is slightly more efficient and is the default storage device for
libvirt.

This is safe to add before systemd is fully integrated, as if the
kernel can't find that file it run /sbin/init, which is busybox