Integrate RemoteArtifactCache into the building process.
For now we do not provide an option to avoid downloading artifacts
during build. Unsetting --cache-server is an option but it's not
user-friendly. Of course if there is no network connection, then
downloading will fail and things will be built locally. We may
want to add a generic --no-network switch to replace/extend
--no-git-update later.