Defer installing chunk artifacts until we need them.
We want to avoid installing anything into the staging area unless
we actually need it. When building we now simply defer all install
operations until we come across something we do not have cached
and that we actually have to build. We then install everything we
have encountered so far (either cached or built) and clear the list
of what needs to be installed.

The list of things to be installed in the future is populated at
the end of processing a build group so that nothing built in the
current group is installed until all other artifacts in the group
have been built as well.