morphlib: add 'arch' field to morphologies
This is an ugly, ugly way to do this, but time is pressing.

SystemBuilder checks what arch is defined in the morphology, if
it is an x86 (or None for compatibility) then it will do the syslinux
install stuff.

This hack is needed because syslinux is x86 specific and arm
often has different requirements for where the kernel must be loaded
from, sometimes it is flash, sometimes it is a different partition.

This will likely become board specific, but for a qemu-system-arm,
the kernel should be a separate file, to be passed on the command line.

Having a different 'kind' for each architecture would be a nicer way,
but would require more changes, since there are various checks for
morphology['kind'] == 'system'