Change install_artifacts to fetch missing artifacts
This is a behavioral change, and is useful when building and having access
to an artifact cache server. Previously, nothing was actually fetching
artifacts from the remote server when the staging area was being crafted.

Morph could do with a bit of cleanup where we separate more cleanly the
acts of fetching stuff to the local artifact cache from the remote one,
when possible, and then using artifacts from the local cache. Currently
this stuff is spread around a bit too much. I did not do that in this
patch, however, to keep undiscussed changes small.