Create class for build logic and fix the logic
The build logic was spread all across the Morph class. This patch
gathers it into one class. Having it in one place is useful for
general code quality reasons, but also makes it feasible to experiment
with the build logic, and override only parts of it.

This patch also changes how the build logic works. Earlier we kept
one staging area for staging builds, and were unpacking everything in
a build group at a time. This meant that the staging area would contain
things that were not strictly build dependencies. This has a possiblity
of changing what gets built. The new logic creates a new, fresh, clean
staging area for each build.

This makes building a bit slower, but we'll fix that by making it really
fast to create a staging area and filling it with build dependencies.