baserock-bootstrap: remove update-gits loop
This should no longer be needed, since gitano on our own server is
more stable and faster than gitorious.org was for us.

Ideally we would have update-gits have a set of repositories to
update and if one fails, increment the fail count and move on to
the next if it hasn't exceeded the retries limit.