Move stratum verification to morphologyfactory
I think morphologies should be verified closer to where they are
created, not when they are traversed, since this verification would
be useful in more code paths.

This is in morphologyfactory, rather than Morphology itself, since
it may be useful to be able to create morphologies without checking,
such as in unit tests.