tests/make-patch: mount disk image for test
Previously it would extract the stratum tarballs, since it was
easier than mounting the disk image.

Then when strata became chunk lists, this created the tarball, then
extracted it.

Now systems may contain files that aren't in the component strata,
this causes problems for tbdiff-deploy, since it tries to alter a
file that isn't there.

So to fix this, it mounts the disk image, like it should have from
the beginning.