Add scripts/clean-artifact-cache
This removes all but the latest build of a chunk from the cache,
which can free up a significant amount of disk space for large
chunks.

Ideally Morph should handle cleaning up the artifact cache itself,
there are some ideas floating around including using git to store
the artifacts; either way, reference counting items in the cache
and only requiring the user to manually delete the final system
images when they are no longer required is probably the best way.