Add test for petrifying strata
This test checks if the same morphology json is generated during
petrification.

Because commits include the author and date, the sha1 sum will
differ if a different user commits at a different time.

This can be overridden by the environment variables
GIT_{AUTHOR,COMMITTER}_{NAME,EMAIL,DATE}.

The date chosen was just something recent at the time the work was
done, since a too-early timestamp is rejected by git as invalid.