SyslinuxDiskBuilder: sync before snapshotting
This is needed in older versions of btrfs, otherwise snapshots can
contain empty files, since the data has not been written to disk.

This behaviour was discovered in a debian squeeze i386 VM, and has not
been seen elsewhere, but this is an environment that must be supported
for bootstrapping.