Rename the "morph minedir" command to "morph workspace"
Mine may be a nice metaphor but it is not very intuitive. Hence, this
commit renames "mine" to "workspace". This affects not just the branch
and merge plugin (the only code that changes) but also documentation
and tests, which are updated along with the code in this commit.

I decided to omit the "dir" suffix in the command because I think
people know well that a workspace will be a directory in this case
and that allows us to refer to this concept with one word ("workspace")
instead of two words ("workspace directory").