Fix branch/checkout to always create branches in the workspace directory
Doing this rather than creating the system branches in the current working
directory allows "morph branch" and "morph checkout" to be run anywhere in
the workspace (e.g. in a different branch).

This commit also adds two tests to verify that new branches are always
created in the toplevel workspace directory.