Warning instead of error for empty strata, temporarily
This kind of reverts commit: c4facff97b27968f00a75ffb3d49cea2080ad9a4

Currently our bootstrap depends on building an empty stratum; this
should be changed so that it builds a tarball system image, but this
is currently broken too. This commit should be reverted when bootstrap
is fixed to not depend on building an empty stratum.