Rewrite edit command to expect SYSTEM STRATUM [CHUNK] parameters
Editing no longer requires a repository to be specified, neither does
it require a branch.

It now starts off from a system morphology that is required to exist
in the branch root repository. Relative to this system, "morph edit"
realises the repository of a stratum and, optionally, a chunk, creates
edit branches named after the system branch, if necessary, and update
the references in the system and stratum morphology accordingly.

The changes made to any of the repositories in the system branch
are not committed.

All existing changes are updated to work with this new input syntax
for "morph edit".