Make "morph edit" print the changes for the user, rework static methods
This commit introduces a new way of logging changes made to
repositories or morphologies as part of executing a morph branch plugin
command. These changes can then logged to the user grouped by
repositories.

Some of the static methods in the branch and merge plugin are converted
to non-static methods.