Return tree SHA1 when looking for ref resolution.
Morph now expects the tree SHA1 in addition when resolving references
using the cache server.  This is to better facilitate correct cache
key computation since commits can be made which have no tree changes
and thus nothing to usefully affect the build.  (For example the morph
branch and build features)