Rename 'morph build' command to 'morph build-morphology'
This is needed because we are replacing the current build command
by a new one that builds from a local system branch. Since that will
be the more common scenario, we want the new implementaiton to go by
the name 'morph build'.

This commit therefore renames 'morph build' and updates all the
tests to use 'morph build-morphology' instead.