Add a /list method
When --enable-writes is set, we provide a /list target which produces
a JSON dictionary of information about the state of the artifact cache.

The dictionary is of the form:

{
    "freespace": NBYTES_OF_SPACE,
    "files": {
        "artifact-filename": {
            "atime": ATIME_AS_NUMBER,
            "size": NBYTES_SIZE_OF_FILE,
            "used": NBYTES_USED_ON_DISK
        },
        ...
    }
}

This allows a controller to decide which artifacts have not been requested in
some time and also how big artifacts are, not only in terms of their 'byte'
size, but also the space they consume on disk.  System images in particular may
differ in this respect since they should be sparsely stored.