Support for fetching artifacts to the cache
Rather than pushing artifacts to the cache, this method allows the caller to
specify a host and artifact which the cache server will then fetch into its
local cache.  It takes the following arguments:

    host=hostname:port
    artifact=artifactname

This is transformed into a fetch to:

    http://hostname:port/artifacts?basename=artifactname

Which is then fetched into the cache under the given name.

The return from this is a JSON object of the form:

{
    "filename": artifactname,
    "size": NBYTES_SIZE_OF_FILE,
    "used": NBYTES_DISK_SPACE_USED
}