Add facility to delete artifacts
In order to allow the artifact cache to be cleaned up, this patch allows for a
/delete method which can remove artifacts from the cache.  It takes the
following arguments:

    artifact=artifactname

The artifact will be deleted and a JSON object returned in the form:

{
    "status": errno,
    "reason": strerror
}

Where errno is zero on success, 1 on EPERM, 2 on ENOENT etc. and reason is the
strerror of the errno, in case the architectures differ between caller and
cache.