Add support for Trove hosted content to morph configuration.
Trove has a fixed structure for repositories and prefixes.  By taking advantage
of this, we can simplify the repo-alias configuration in morph somewhat and
thus make it easier for others to add their own aliases and prefixes later,
without worrying that they'll miss out on default ones created by the Baserock
team.

We override process_args in the application class to achieve this because
cliapp's setup() hook is not quite right for our purposes.