Test the trove-* support.
This adds a way to cause the processed configuration to be dumped out.  We do
this in the overridden process_args method so that the configuration dumped has
had all the processing done on it which is necessary.

The test then uses this functionality to ensure that the repo-alias setting has
been correctly updated with the Trove-hosted prefixes.