morph merge: Merge changes in morphology contents
We now load the previous version of the morphology and check for
differences, which allows us to handle components that were added or
removed between branches.

We currently disallow adding new chunks or strata in a branch if
they were 'edit'ed afterwards. We can fix this by having 'edit'
store the original ref in the repo, rather than reading it from
the old version of the morphology.