Rework git caches to be bare mirrors of the repos.
This reworks the code for managing and using the git caches in morph to treat
the caches as bare repositories which are mirrors of where we clone from.  In
addition we correctly prune the branches during updates, so that we don't end
up accumulating pointless branches over and over.  This is even more important
with branch-and-merge generating temporary build refs for things.