Always cache artifacts locally and report filenames during build.
The artifact being built is often subsequently hard to find since the user has
to go digging around in their artifact cache to try and determine which
artifact is the one they want.

With this change, the filepath to the artifact will be displayed after building
it.  As a side effect, we also need to always cache artifacts locally in order
to do this, otherwise we'd have to report URLs to the remote artifact cache
instead and that might be less useful overall.