Use repository cache for git access other than current system branch
The origin/ refs in the system branch checkout repos may or may not be
up to date, and may or may not have been tampered with by the user.
Much better to use our central cache for everything other than changes
to the system branch ref itself, where we should indeed be honouring
the user's local changes.

At a later date we could warn if the user modifies refs other than
the system branch ref but does not push, as these changes will have no
effect.

NOTE: this commit breaks 'morph merge'. It is fixed in the next commit.