Change the time sanity check
We were having trouble with NTP setting the system time to be correct
after .pyc files had already been created. The .pyc files got created
at startup, before NTP had had a chance to run. This sometimes resulted
in .pyc timestamps being wildly wrong and far in the future or past,
and that would prevent morph from ever working again on the machine.

We now just check that the system time is not far in the past. This
is enough to make sure we don't create timestamps that break tar.

Reviewed-By: Daniel Silverstone (on irc)