morph merge: Warn if a merge will cause unexpected petrification
The concept of a component path is new. This is simply a concise way
of referring to a component in an error message, and looks like this:

    base-system-x86_64-generic.bsp-x86_64-generic.linux

We currently only touch the 'edited chunks' in merge_stratum(), i.e.
those in the FROM branch where 'morph edit' was run. However, the
petrification can affect any chunk so there is a new method added to
obtain all components in a morphology. This function also returns the
differences between the two, which we will make use of at a later date.